This commit is important/ Final Working file.
I am committing this file as it has been my working file the whole
time. It started as a new project in web storm and was just copied into
github. It does not have a js folder either. What makes it important is
that it works FINE! I submitted both to show what I am seeing from my
side.